OCT-111: moved IsCatalogsNumberLimitReachedQuery
